+++ app/admin/course_schedules.rb
@@ -27,16 +27,14 @@ ActiveAdmin.register CourseSchedule do
     end
   end
 
-  scope I18n.t('activerecord.attributes.course_schedule.states.active')    , :active, default: true
-  scope I18n.t('activerecord.attributes.course_schedule.states.suspended') , :suspended
+  scope I18n.t('activerecord.attributes.course_schedule.states.active'), :active, default: true
+  scope I18n.t('activerecord.attributes.course_schedule.states.suspended'), :suspended
   scope I18n.t('ui.users.commons.all')                        , :all
 
   index do
     selectable_column
     id_column
-    if current_admin_user.is_root?
-      column(:organization)
-    end
+    column(:organization) if current_admin_user.is_root?
     column(:category)
     column(:course)
     column(:room)