+++ app/jobs/update_course_events_job.rb
@@ -1,7 +1,7 @@
 class UpdateCourseEventsJob < ApplicationJob
   queue_as :course_events
 
-  def perform(*args)
+  def perform
     CourseEvent.where(state: :active).where('event_date < ?', Time.zone.now).update_all state: :closed
   end
 end