+++ config/application.rb
@@ -24,8 +24,8 @@ module Insiemento
 
     # config.assets.precompile += %w[active_admin.scss active_admin.js]
 
-    #config.active_job.queue_adapter = :sidekiq
-    #config.action_mailer.deliver_later_queue_name = "default_mailer"
+    config.active_job.queue_adapter = :sidekiq
+    config.action_mailer.deliver_later_queue_name = "default_mailer"
 
     config.active_storage.variant_processor = :vips