+++ db/migrate/20200827131835_create_user_document_models.rb
@@ -1,0 +1,13 @@
+class CreateUserDocumentModels < ActiveRecord::Migration[6.0]
+  def change
+    create_table :user_document_models do |t|
+      t.references :organization, null: false, foreign_key: true
+      t.integer :state, limit: 1
+      t.string :title, limit: 255
+      t.text :body
+      t.integer :validity_days, limit: 2
+
+      t.timestamps
+    end
+  end
+end