+++ spec/factories/user_documents.rb
@@ -1,0 +1,8 @@
+FactoryBot.define do
+  factory :user_document do
+    organization { nil }
+    user_document_model { nil }
+    user { nil }
+    state { :new }
+  end
+end